+++ src/main/java/nextstep/auth/AuthService.java
@@ -1,5 +1,6 @@
 package nextstep.auth;
 
+import nextstep.exception.AuthenticationException;
 import nextstep.member.Member;
 import nextstep.member.MemberDao;
 import org.springframework.stereotype.Service;
