+++ src/main/java/nextstep/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package nextstep.exception;
 
 public class AuthenticationException extends RuntimeException {
