+++ src/main/java/nextstep/exception/NotExistEntityException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class NotExistEntityException extends RuntimeException {
 }
