+++ src/main/java/nextstep/exception/UnAuthorizationException.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package nextstep.exception;
 
 public class UnAuthorizationException extends RuntimeException {
 }
